split table index, rounded table
