+++ src/assets-webkit/styles/serlo-tailwind.css
@@ -61,13 +61,13 @@
     @apply mx-side mt-0 mb-5;
   }
   .serlo-table {
-    @apply mx-side border-collapse whitespace-nowrap break-normal special-hyphens-auto mobile:whitespace-normal;
+    @apply mx-side border-separate border-spacing-0 whitespace-nowrap break-normal border-brand-200 special-hyphens-auto mobile:whitespace-normal;
   }
   .serlo-th {
-    @apply border-3 border-brand-200 bg-brand-100 p-1.5 pl-3 text-left;
+    @apply border-r-3 border-b-3 border-brand-200 bg-brand-100 p-1.5 pl-3 text-left;
   }
   .serlo-td {
-    @apply border-3 border-brand-200 p-1.5 align-top;
+    @apply border-r-3 border-b-3 border-brand-200 p-1.5 align-top;
     > p.serlo-p {
       @apply mx-0;
     }
